\name{RIJ2-package}
\alias{RIJ2-package}
\alias{RIJ2}
\docType{package}
\title{
\packageTitle{RIJ2}
}
\description{
\packageDescription{RIJ2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RIJ2}
\packageIndices{RIJ2}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RIJ2}

Maintainer: \packageMaintainer{RIJ2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
